requirements.txt
astropy
codecov
extinction
fastavro
fink-client
matplotlib
numpy
pandas
pyyaml
scipy
sncosmo


